requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
websockets==15.0.1
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-multipart>=0.0.6
email-validator==2.3.0
cryptography>=42.0.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1
SQLAlchemy==2.0.45
alembic==1.17.2
psycopg2-binary==2.9.11
PyMySQL==1.1.1
requests==2.32.5
pandas>=2.0.0,<2.3.0
openpyxl>=3.0.0
xlrd>=2.0.0
numpy>=1.24.0
gunicorn
websockets==15.0.1